05-Reinforcement-Learning-Noise-Mitigation/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "qrl-noise-mitigation"
version = "0.1.0"
description = "Research benchmark for measurement-defined quantum RL under NISQ noise"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "numpy==2.4.3",
  "matplotlib==3.10.8",
  "PyYAML==6.0.3",
  "pandas==3.0.1",
  "jupyter==1.1.1",
  "qiskit==2.3.1",
  "qiskit-aer==0.17.2",
  "qiskit-ibm-runtime==0.45.1",
  "qiskit-machine-learning==0.9.0",
]

[project.scripts]
qrl-train = "src.training_pipeline:main"
qrl-sweep = "src.hyperparameter_sweep:main"
qrl-audit = "src.experiment_design_audit:main"
qrl-benchmark = "src.benchmark_suite:main"
qrl-report = "src.paper_report:main"
qrl-mitigation-ablation = "src.mitigation_ablation_suite:main"
qrl-environment-sweep = "src.environment_sweep:main"
qrl-hardware-audit = "src.hardware_audit:main"
qrl-hardware-eval = "src.hardware_evaluation:main"

[tool.setuptools.packages.find]
include = ["agent*", "src*", "utils*", "environments*", "core*", "methods*", "benchmarks*", "reporting*", "hardware*"]

[tool.setuptools]
include-package-data = true
